requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv
Pillow==10.2.0
